+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+  - basic testing suite
+  - `compilerDidMount` event
 
 ## [0.5.0] - 2017-12-20
 ### Added